+++ src/main/java/com/ssg/usms/business/Security/login/UsmsUserDetails.java
@@ -1,4 +1,4 @@
-package com.ssg.usms.business.security.login;
+package com.ssg.usms.business.Security.login;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.ssg.usms.business.user.dto.SecurityState;
