+++ src/main/java/com/ssg/usms/business/Security/logout/logoutSuccessHandler.java
@@ -1,7 +1,7 @@
-package com.ssg.usms.business.security.logout;
+package com.ssg.usms.business.Security.logout;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ssg.usms.business.security.login.persistence.ResponseLoginDto;
+import com.ssg.usms.business.Security.login.persistence.ResponseLoginDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
