+++ src/main/java/com/ssg/usms/business/config/SecurityConfig.java
@@ -1,9 +1,9 @@
 package com.ssg.usms.business.config;
 
 
-import com.ssg.usms.business.security.login.UsmsLoginConfiguer;
-import com.ssg.usms.business.security.login.authority.UsmsAccessDeniedHandler;
-import com.ssg.usms.business.security.login.authority.UsmsForbiddenEntryPoint;
+import com.ssg.usms.business.Security.login.UsmsLoginConfiguer;
+import com.ssg.usms.business.Security.login.authority.UsmsAccessDeniedHandler;
+import com.ssg.usms.business.Security.login.authority.UsmsForbiddenEntryPoint;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
