+++ src/test/java/com/ssg/usms/business/Security/login/LoginIntegrationTest.java
@@ -1,8 +1,8 @@
-package com.ssg.usms.business.security.login;
+package com.ssg.usms.business.Security.login;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ssg.usms.business.security.login.persistence.RequestLoginDto;
-import com.ssg.usms.business.security.login.persistence.ResponseLoginDto;
+import com.ssg.usms.business.Security.login.persistence.RequestLoginDto;
+import com.ssg.usms.business.Security.login.persistence.ResponseLoginDto;
 import com.ssg.usms.business.user.repository.UserRepository;
 import com.ssg.usms.business.user.repository.UsmsUser;
 import lombok.extern.slf4j.Slf4j;
